also adding map
