+++ extension/core_functions/scalar/map/functions.json
@@ -7,13 +7,6 @@
         "type": "scalar_function"
     },
     {
-        "name": "map",
-        "parameters": "keys,values",
-        "description": "Creates a map from a set of keys and values",
-        "example": "map(['key1', 'key2'], ['val1', 'val2'])",
-        "type": "scalar_function_set"
-    },
-    {
         "name": "map_entries",
         "parameters": "map",
         "description": "Returns the map entries as a list of keys/values",
