+++ src/function/function_list.cpp
@@ -142,6 +142,7 @@ static const StaticFunctionDefinition function[] = {
 	DUCKDB_SCALAR_FUNCTION(ListZipFun),
 	DUCKDB_SCALAR_FUNCTION(LowerFun),
 	DUCKDB_SCALAR_FUNCTION_SET(LtrimFun),
+	DUCKDB_SCALAR_FUNCTION_SET(MapFun),
 	DUCKDB_SCALAR_FUNCTION(MapContainsFun),
 	DUCKDB_AGGREGATE_FUNCTION_SET(MaxFun),
 	DUCKDB_SCALAR_FUNCTION_SET(MD5Fun),
