+++ src/function/scalar/map/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library_unity(duckdb_function_map OBJECT map_contains.cpp)
+add_library_unity(duckdb_function_map OBJECT map_contains.cpp map.cpp)
 
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:duckdb_function_map>
