+++ src/function/scalar/map/functions.json
@@ -5,5 +5,12 @@
         "description": "Checks if a map contains a given key.",
         "example": "map_contains(MAP {'key1': 10, 'key2': 20, 'key3': 30}, 'key2')",
         "type": "scalar_function"
+    },
+    {
+        "name": "map",
+        "parameters": "keys,values",
+        "description": "Creates a map from a set of keys and values",
+        "example": "map(['key1', 'key2'], ['val1', 'val2'])",
+        "type": "scalar_function_set"
     }
 ]
