+++ src/function/scalar/map/map.cpp
@@ -1,4 +1,4 @@
-#include "core_functions/scalar/map_functions.hpp"
+#include "duckdb/function/scalar/map_functions.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
 #include "duckdb/common/string_util.hpp"
 #include "duckdb/parser/expression/bound_expression.hpp"
