+++ src/include/duckdb/function/scalar/map_functions.hpp
@@ -25,4 +25,14 @@ struct MapContainsFun {
 	static ScalarFunction GetFunction();
 };
 
+struct MapFun {
+	static constexpr const char *Name = "map";
+	static constexpr const char *Parameters = "keys,values";
+	static constexpr const char *Description = "Creates a map from a set of keys and values";
+	static constexpr const char *Example = "map(['key1', 'key2'], ['val1', 'val2'])";
+	static constexpr const char *Categories = "";
+
+	static ScalarFunctionSet GetFunctions();
+};
+
 } // namespace duckdb
